core: fix transaction test case (cherry-pick #15021) (#15070)
* core: fix transaction test case (#15021)

* move patched ct to root


* use our transaction test case as base


* fix...?


* well apparently that works


---------


* fixup! core: fix transaction test case (#15021)

---------
